fix mobile copy and scroll issues
