+++ frontend/components/ui/CopyButton.tsx
@@ -2,6 +2,7 @@
 
 import { useState, useCallback } from 'react';
 import { Check, Copy } from 'lucide-react';
+import { copyText } from '@/lib/copyText';
 
 interface CopyButtonProps {
   /** Text that will be copied to the clipboard */
@@ -13,23 +14,7 @@ export default function CopyButton({ code }: CopyButtonProps) {
 
   // Copy provided code snippet to the clipboard
   const handleCopy = useCallback(async () => {
-    try {
-      await navigator.clipboard.writeText(code);
-    } catch {
-      // Fallback for older mobile browsers
-      const textarea = document.createElement('textarea');
-      textarea.value = code;
-      textarea.style.position = 'fixed';
-      textarea.style.opacity = '0';
-      document.body.appendChild(textarea);
-      textarea.focus();
-      textarea.select();
-      try {
-        document.execCommand('copy');
-      } finally {
-        document.body.removeChild(textarea);
-      }
-    }
+    await copyText(code);
     setCopied(true);
     setTimeout(() => setCopied(false), 2000);
   }, [code]);
